+++ packages/spacecat-shared-utils/src/functions.js
@@ -15,6 +15,16 @@ const REGEX_ISO_DATE = /\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z/;
 const REGEX_TIME_OFFSET_DATE = /\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}(Z|[+-]\d{2}:\d{2})/;
 
 /**
+ * Determines if the given parameter is an array.
+ *
+ * @param {*} value - The value to check.
+ * @returns {boolean} True if the parameter is an array, false otherwise.
+ */
+function isArray(value) {
+  return Array.isArray(value);
+}
+
+/**
  * Determines case-insensitively if the given value is a boolean or a string
  * representation of a boolean.
  *
@@ -53,7 +63,7 @@ function isNumber(value) {
  * @returns {boolean} True if the parameter is an object, false otherwise.
  */
 function isObject(value) {
-  return !Array.isArray(value) && value !== null && typeof value === 'object';
+  return !isArray(value) && value !== null && typeof value === 'object';
 }
 
 /**
@@ -150,14 +160,15 @@ function toBoolean(value) {
  * @param {Array} b - The second array to compare.
  * @returns {boolean} True if the arrays are equal, false otherwise.
  */
-const arrayEquals = (a, b) => Array.isArray(a)
-  && Array.isArray(b)
+const arrayEquals = (a, b) => isArray(a)
+  && isArray(b)
   && a.length === b.length
   && a.every((val, index) => val === b[index]);
 
 export {
   arrayEquals,
   hasText,
+  isArray,
   isBoolean,
   isInteger,
   isValidDate,